refactor(factoryconfig): split guard and resource mapping
