+++ docs/internal/baselines/deadcode-baseline.txt
@@ -171,7 +171,7 @@ pkg/services/factory_definitions/transports/mapping/factoryconfig/diagnostics/di
 pkg/services/factory_definitions/transports/mapping/factoryconfig/diagnostics/diagnostics.go: unreachable func: ValidationResult.Errors
 pkg/services/factory_definitions/transports/mapping/factoryconfig/diagnostics/diagnostics.go: unreachable func: ValidationResult.HasErrors
 pkg/services/factory_definitions/transports/mapping/factoryconfig/diagnostics/diagnostics.go: unreachable func: canonicalTopologyFindings
-pkg/transports/mapping/factoryconfig/factory_config_mapping_internal.go: unreachable func: cloneFactoryOrchestratorConfig
+pkg/transports/mapping/factoryconfig/invocation_examples_compatibility.go: unreachable func: cloneFactoryOrchestratorConfig
 pkg/transports/mapping/factorysession/factory_session_mapper_test.go: unreachable func: fixtureDispatchRows
 pkg/transports/mapping/factorysession/factory_session_projection_test.go: unreachable func: resultRequestFromFixture
 pkg/transports/mapping/session_api.go: unreachable func: ComposeSessionAPI
